Fix `html` links in the all students overview
When running stacscheck across all students, the overview's `html` link
for each student erroneously points to the `txt/<MatricNo>.html`, which
is incorrect. Other members of staff seem to either ignore it, work
around it by clicking on an individual test (which links correctly), or
some might have their own ways of fixing it locally. In any case, this
commit fixes the link generation to point to `html/<MatricNo>.html`,
which should save a small amount of friction in future runs.